requirements.txt
arabic-reshaper==2.1.3
asgiref==3.4.1
certifi==2021.10.8
charset-normalizer==2.0.7
cssselect2==0.4.1
Django==3.2.9
django-cors-headers==3.10.0
django-filter==21.1
djangorestframework==3.12.4
djangorestframework-simplejwt==5.0.0
easy-thumbnails==2.8
future==0.18.2
html5lib==1.1
idna==3.3
lxml==4.7.1
packaging==21.3
Pillow==9.0.0
PyJWT==2.3.0
pyparsing==3.0.6
PyPDF2==1.26.0
python-bidi==0.4.2
python-dotenv==0.19.2
python-magic==0.4.24
pytz==2021.3
razorpay==1.3.0
reportlab==3.6.5
requests==2.26.0
six==1.16.0
sqlparse==0.4.2
svglib==1.1.0
tinycss2==1.1.1
urllib3==1.26.7
webencodings==0.5.1
xhtml2pdf==0.2.5
